requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlmodel>=0.0.16
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.0.0
httpx>=0.27.0
